test(lists): consolidate hosts.go helper tests into one file
Move the plain table/fuzz tests for MightBeIP, isValidDomainName and
normalizeHostsListEntry out of three separate files into a single
hosts_funcs_test.go that pairs with hosts.go. They stay separate from the
Ginkgo behaviour suite (hosts_test.go) because Fuzz targets cannot be written
as Ginkgo It blocks. No test logic changes.
